unifying methods to `WebbBridge` contract
